requirements.txt
discord.py
python-dotenv
requests
backports-datetime-fromisoformat